+++ CHANGES.rst
@@ -20,6 +20,7 @@ in progress
 - Upgrade to webpack4
 - Improve auto layouter significantly
 - Improve PDF rendering subsystem significantly
+- Use specific nodejs version for decktape
 
 
 2018-02-08 0.4.0